+++ src/main/java/com/mindhub/homebanking/repositories/AccountRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface AccountRepo extends JpaRepository<Account, Integer> {
+public interface AccountRepo extends JpaRepository<Account, Long> {
 }
