+++ src/main/java/com/mindhub/homebanking/repositories/CardRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface CardRepo extends JpaRepository<Card, Integer> {
+public interface CardRepo extends JpaRepository<Card, Long> {
 }
