+++ src/main/java/com/mindhub/homebanking/repositories/LoanRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface LoanRepo extends JpaRepository<Loan, Integer> {
+public interface LoanRepo extends JpaRepository<Loan, Long> {
 }
