+++ src/main/java/com/mindhub/homebanking/repositories/TransactionRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface TransactionRepo extends JpaRepository<Transaction, Integer> {
+public interface TransactionRepo extends JpaRepository<Transaction, Long> {
 }
